support multiple text outputs and ordering
