+++ examples/next/src/app/page.tsx
@@ -8,9 +8,8 @@ export default function Home() {
       <div className="relative flex w-full h-full overflow-hidden">
         <div className="group w-full pl-0 animate-in duration-300 ease-in-out peer-[[data-state=open]]:lg:pl-[250px] peer-[[data-state=open]]:xl:pl-[300px]">
           <OpenAssistantsChat
-            api={`${
-              process.env.NEXT_PUBLIC_API_URL || 'http://localhost:8000'
-            }/v1alpha/assistants/hooli/chat`}
+            api={`${process.env.NEXT_PUBLIC_API_URL ||
+              'http://localhost:8000'}/v1alpha/assistants/hooli/chat`}
             name="Example Assistant"
             starterPrompts={[
               {
@@ -18,8 +17,8 @@ export default function Home() {
                 prompt: 'Show recent purchases',
               },
               {
-                title: 'Show spend per product',
-                prompt: 'Show spend per product',
+                title: 'Show spend all products',
+                prompt: 'Show spend all products',
               },
               {
                 title: 'What can I ask?',
