+++ packages/openassistants-react/src/components/chat-list.tsx
@@ -33,62 +33,17 @@ export function OpenAssistantsChatList({
   return (
     <div className="relative mx-auto max-w-2xl px-4">
       {messages.map((message, index) => {
-        const outputs = 'outputs' in message ? message.outputs : null;
         return (
           <div key={index}>
-            {outputs &&
-              outputs.map((output, i) => {
-                console.log(
-                  `output.type=${output.type} output=${JSON.stringify(output)}`
-                );
-                return (
-                  <div key={`message_${index}_output_${i}`} className="mb-4">
-                    {output.type === 'dataframe' && (
-                      <DataframeTable
-                        key={`message_${index}_df_${i}`}
-                        title={output.title || ''}
-                        dataframe={output.dataframe}
-                      ></DataframeTable>
-                    )}
-                    {output.type === 'visualization' && (
-                      <RenderVisMessage
-                        id={`message-${index}-output-${i}`}
-                        data={output.visualization.data}
-                        layout={output.visualization.layout}
-                      />
-                    )}
-                  </div>
-                );
-              })}
-
-            {getContent(message) && (
-              <div key={index}>
-                <OpenAssistantsChatMessage
-                  key={index}
-                  message={message}
-                  onEdit={(m) => onEdit(index, m)}
-                  isLoading={isLoading}
-                />
-                {index < messages.length - 1 && <Separator className="my-4 " />}
-              </div>
-            )}
-
-            {message.role === 'assistant' && message.input_request && (
-              <FunctionForm
-                message={message}
-                onSubmit={(values) => {
-                  const newMessage = {
-                    role: 'user' as const,
-                    input_response: {
-                      name: message.input_request!.name,
-                      arguments: values,
-                    },
-                    content: '',
-                  };
-                  append(newMessage);
-                }}
-              ></FunctionForm>
-            )}
+            <OpenAssistantsChatMessage
+              key={index}
+              total={messages.length}
+              index={index}
+              message={message}
+              append={append}
+              onEdit={(m) => onEdit(index, m)}
+              isLoading={isLoading}
+            />
           </div>
         );
       })}
